refactor : 템플릿, 이벤트 핸들러, 유틸성 메소드 분리
