+++ assets/javascript/template/post.js
@@ -1,4 +1,6 @@
 import { postData } from "../data/posts.js";
+import { insertBeforeEnd } from "../util.js";
+import { addPostEventListener } from "../eventHandler/postEventHandler.js";
 
 export function postTemplate(data) {
   const {
@@ -17,19 +19,17 @@ export function postTemplate(data) {
   <article class="post">
     <div class="post_header">
       <div class="author_img">
-        <!-- TODO: alt 변경-->
-        <img src="${authorSrc}" alt="">
+        <img src="${authorSrc}" alt="author image">
       </div>
       <div class="author_info">
         <div class="author_name">${authorName}</div>
         <div class="author_location">${authorLocation}</div>
       </div>
-      <!-- TODO: button 변경-->
+      <!-- TODO: button 작동하도록 변경-->
       <div class="more_btn">...</div>
     </div>
     <div class="post_img">
-      <!-- TODO: alt 변경-->
-      <img src="${postSrc}" alt="199010">
+      <img src="${postSrc}" alt="post image">
     </div>
     <div class="post_reaction">
       <div class="reaction_btns">
@@ -51,8 +51,7 @@ export function postTemplate(data) {
       </div>
     </div>
     <div class="post_likes">
-      <!-- TODO: alt 변경-->
-      <div class="likes_thumbnail"><img src="${likeThumbnail}" alt=""></div>
+      <div class="likes_thumbnail"><img src="${likeThumbnail}" alt="thumbnail of liker representative"></div>
       <div class="likes_info"><b>${likeName}</b> 님 <b>외 <span class="like_count_${index}">${likeCount}</span>명</b>이 좋아합니다.</div>
     </div>
     <div class="post_body">
@@ -72,26 +71,22 @@ export function postTemplate(data) {
   </article>`
 }
 
-function likeButton() {
-  return `
+export const likeButton = `
     <svg aria-label="좋아요" class="_8-yf5 " fill="#262626" height="24" viewBox="0 0 48 48" width="24">
       <path clip-rule="evenodd"
           d="M34.3 3.5C27.2 3.5 24 8.8 24 8.8s-3.2-5.3-10.3-5.3C6.4 3.5.5 9.9.5 17.8s6.1 12.4 12.2 17.8c9.2 8.2 9.8 8.9 11.3 8.9s2.1-.7 11.3-8.9c6.2-5.5 12.2-10 12.2-17.8 0-7.9-5.9-14.3-13.2-14.3zm-1 29.8c-5.4 4.8-8.3 7.5-9.3 8.1-1-.7-4.6-3.9-9.3-8.1-5.5-4.9-11.2-9-11.2-15.6 0-6.2 4.6-11.3 10.2-11.3 4.1 0 6.3 2 7.9 4.2 3.6 5.1 1.2 5.1 4.8 0 1.6-2.2 3.8-4.2 7.9-4.2 5.6 0 10.2 5.1 10.2 11.3 0 6.7-5.7 10.8-11.2 15.6z"
           fill-rule="evenodd"></path>
     </svg>
          
-  `
-}
+  `;
 
-function unlikeButton() {
-  return `
+export const unlikeButton = `
     <svg aria-label="좋아요 취소" class="_8-yf5 " fill="#ed4956" height="24" viewBox="0 0 48 48" width="24">
       <path clip-rule="evenodd" d="M35.3 35.6c-9.2 8.2-9.8 8.9-11.3 8.9s-2.1-.7-11.3-8.9C6.5 30.1.5 25.6.5 17.8.5 9.9 6.4 3.5 13.7 3.5 20.8 3.5 24 8.8 24 8.8s3.2-5.3 10.3-5.3c7.3 0 13.2 6.4 13.2 14.3 0 7.8-6.1 12.3-12.2 17.8z" fill-rule="evenodd"></path>
     </svg>
-  `
-}
+  `;
 
-function comment(commentContent) {
+export function comment(commentContent) {
   return `
     <div class="comment">
       <div class="author_name">woowa.crew</div>
@@ -100,85 +95,8 @@ function comment(commentContent) {
   `
 }
 
-function toggleLike(index) {
-  const like = document.querySelector(`.btn_${index}`);
-  const likeCount = document.querySelector(`.like_count_${index}`);
-  const currentLikeCount = likeCount.innerText;
-  if (like.innerHTML === unlikeButton()) {
-    like.innerHTML = likeButton();
-    likeCount.innerHTML = parseInt(currentLikeCount) - 1;
-    return;
-  }
-  like.innerHTML = unlikeButton();
-  likeCount.innerHTML = parseInt(currentLikeCount) + 1;
-}
-
-function handleCommentInputDisplay(index) {
-  const commentButton = document.querySelector(`.btn_comment_${index}`)
-  const commentInputBox = document.querySelector(`.comment_${index}`);
-  commentButton.addEventListener("click", function () {
-    console.log(commentInputBox.style.display);
-    commentInputBox.style.display = commentInputBox.style.display !== "flex" ? "flex" : "none";
-  })
-}
-
-function handleCommentInput(index) {
-  const input = document.querySelector(`.comment_input_${index}`);
-  const button = document.querySelector(`.comment_input_submit_${index}`);
-  input.addEventListener("input", function (e) {
-        e.target.value.trim().length > 0 ?
-            button.removeAttribute("disabled")
-            : button.setAttribute("disabled", "true");
-      }
-  )
-}
-
-function stringToHtml(string) {
-  return new DOMParser().parseFromString(string, "text/html").body.firstChild
-}
-
-function addComment(index) {
-  const inputBox = document.querySelector(`.comment_${index}`);
-  const input = document.querySelector(`.comment_input_${index}`);
-  const button = document.querySelector(`.comment_input_submit_${index}`);
-  const comments = document.querySelector(`.comments_${index}`)
-  input.addEventListener("keydown", function (e) {
-        if (e.code === "Enter" && input.value.trim().length > 0) {
-          comments.insertBefore(stringToHtml(comment(input.value)), inputBox);
-          inputBox.style.display = "none";
-          input.value = "";
-          button.setAttribute("disabled", "true");
-        }
-      }
-  )
-  button.addEventListener("click", function () {
-    comments.insertBefore(stringToHtml(comment(input.value)), inputBox);
-    inputBox.style.display = "none";
-    input.value = "";
-    button.setAttribute("disabled", "true");
-  })
-}
-
-function handleLikeToggle(index) {
-  document.querySelector(`.btn_${index}`).addEventListener("click",
-      () => toggleLike(index)
-  )
-}
-
-function mapDataToTemplate() {
-  const posts = document.querySelector(".posts");
-  postData.map(data => postTemplate(data))
-      .forEach(template => {
-        posts.insertAdjacentHTML("beforeend", template);
-      })
-}
 
 export function loadPosts() {
-  mapDataToTemplate();
-  postData.forEach(({index}) => {
-    handleLikeToggle(index);
-    handleCommentInputDisplay(index);
-    handleCommentInput(index);
-    addComment(index)
-  });
+  insertBeforeEnd(".posts", postTemplate, postData)
+  addPostEventListener(postData);
 }
