+++ index.html
@@ -13,12 +13,6 @@
     <link rel="shortcut icon" href="./favicon.ico" type="image/x-icon">
     <link rel="stylesheet" href="./assets/css/style.css">
     <link href="https://fonts.googleapis.com/css?family=Do+Hyeon&display=swap" rel="stylesheet">
-    <script type="module" src="./assets/javascript/main.js"></script>
-    <script type="module" src="./assets/javascript/template/friend.js"></script>
-    <script type="module" src="./assets/javascript/template/header.js"></script>
-    <script type="module" src="./assets/javascript/template/post.js"></script>
-    <script type="module" src="./assets/javascript/data/friends.js"></script>
-    <script type="module" src="./assets/javascript/data/posts.js"></script>
   </head>
   <body>
     <div class="container">
@@ -27,5 +21,13 @@
       <main class="posts"></main>
       <footer class="footer"></footer>
     </div>
+    <script type="module" src="./assets/javascript/main.js"></script>
+    <script type="module" src="./assets/javascript/template/friend.js"></script>
+    <script type="module" src="./assets/javascript/template/header.js"></script>
+    <script type="module" src="./assets/javascript/template/post.js"></script>
+    <script type="module" src="assets/javascript/eventHandler/postEventHandler.js"></script>
+    <script type="module" src="./assets/javascript/data/friends.js"></script>
+    <script type="module" src="./assets/javascript/data/posts.js"></script>
+    <script type="module" src="./assets/javascript/util.js"></script>
   </body>
 </html>
